+++ src/Components/Security/SecurityUtils.hpp
@@ -21,7 +21,7 @@ Security::Strategy ConvertToStrategy(std::string strategy);
 std::unique_ptr<ISecurityStrategy> CreateStrategy(
     Security::Strategy strategy, Security::Role role, Security::Context context);
 
-void EraseMemory(void* begin, std::uint32_t size);
+void EraseMemory(void* begin, std::size_t size);
 
 //------------------------------------------------------------------------------------------------
 } // Security namespace
